move device parser tests
